+++ ioc/iochooks.cpp
@@ -165,9 +165,17 @@ void testCleanupPrepare()
     resetGroups();
 }
 
-////////////////////////////////////
-// Two ioc shell commands for pvxs
-////////////////////////////////////
+static
+void pvxs_log_config(const char *str)
+{
+    logger_config_str(str);
+}
+
+static
+void pvxs_log_reset()
+{
+    logger_level_clear();
+}
 
 /**
  * Show the PVXS server report.
@@ -456,6 +464,12 @@ void pvxsBaseRegistrar() noexcept {
 
         bool enableQ = enable2();
 
+        IOCShCommand<const char*>("pvxs_log_config", "KEY=VAL,KEY=VAL,...",
+                                  "Append logger configuration.  eg. pvxs_log_config \"pvxs.*=DEBUG\"")
+                .implementation<&pvxs_log_config>();
+        IOCShCommand<>("pvxs_log_clear",
+                       "Reset logger configuration to defaults")
+                .implementation<&pvxs_log_reset>();
         IOCShCommand<int>("pvxsr", "[show_detailed_information?]", "PVXS Server Report.  "
                                                                    "Shows information about server config (level==0)\n"
                                                                    "or about connected clients (level>0).\n")