+++ src/clientconn.cpp
@@ -61,9 +61,12 @@ void Connection::startConnecting()
 {
     assert(!this->bev);
 
+    evsocket sock(peerAddr.family(), SOCK_STREAM, 0);
     decltype(this->bev) bev(__FILE__, __LINE__,
-                bufferevent_socket_new(context->tcp_loop.base, -1,
+                bufferevent_socket_new(context->tcp_loop.base, sock.sock,
                                        BEV_OPT_CLOSE_ON_FREE|BEV_OPT_DEFER_CALLBACKS));
+
+    sock.release(); // hand-off ownership to bev
 
     bufferevent_setcb(bev.get(), &bevReadS, nullptr, &bevEventS, this);
 
@@ -381,13 +384,24 @@ void Connection::handle_CREATE_CHANNEL()
     chan->statRx += rxlen;
 
     if(!sts.isSuccess()) {
-        // server refuses to create a channel, but presumably responded positively to search
+        if(chan->forcedServer.family()==AF_UNSPEC) {
+            // server refuses to create a channel, but presumably responded positively to search.
+            // try again
 
-        chan->state = Channel::Searching;
-        context->searchBuckets[context->currentBucket].push_back(chan);
+            log_warn_printf(io, "Server %s refuses channel to '%s' : %s\n", peerName.c_str(),
+                            chan->name.c_str(), sts.msg.c_str());
 
-        log_warn_printf(io, "Server %s refuses channel to '%s' : %s\n", peerName.c_str(),
-                        chan->name.c_str(), sts.msg.c_str());
+            chan->state = Channel::Searching;
+            context->searchBuckets[context->currentBucket].push_back(chan);
+
+        } else {
+            // server refused after we bypassed search, so can't use usual retry method.
+            // refuse to create a tight retry loop, and drop on the floor for now.
+            // retry on reconnect.
+            log_err_printf(io, "Server %s refuses direct channel to '%s' : %s\n", peerName.c_str(),
+                            chan->name.c_str(), sts.msg.c_str());
+            return;
+        }
 
     } else {
         chan->state = Channel::Active;