+++ src/util.cpp
@@ -416,9 +416,7 @@ size_t SockAddr::size() const noexcept
 {
     switch(store.sa.sa_family) {
     case AF_INET: return sizeof(store.in);
-#ifdef AF_INET6
     case AF_INET6: return sizeof(store.in6);
-#endif
     default: // AF_UNSPEC and others
         return sizeof(store);
     }
@@ -428,9 +426,7 @@ unsigned short SockAddr::port() const noexcept
 {
     switch(store.sa.sa_family) {
     case AF_INET: return ntohs(store.in.sin_port);
-#ifdef AF_INET6
     case AF_INET6:return ntohs(store.in6.sin6_port);
-#endif
     default: return 0;
     }
 }
@@ -439,9 +435,7 @@ void SockAddr::setPort(unsigned short port)
 {
     switch(store.sa.sa_family) {
     case AF_INET: store.in.sin_port = htons(port); break;
-#ifdef AF_INET6
     case AF_INET6:store.in6.sin6_port = htons(port); break;
-#endif
     default:
         throw std::logic_error("SockAddr: set family before port");
     }
@@ -559,9 +553,7 @@ bool SockAddr::isAny() const noexcept
 {
     switch(store.sa.sa_family) {
     case AF_INET: return store.in.sin_addr.s_addr==htonl(INADDR_ANY);
-#ifdef AF_INET6
     case AF_INET6: return IN6_IS_ADDR_UNSPECIFIED(&store.in6.sin6_addr);
-#endif
     default: return false;
     }
 }
@@ -570,9 +562,7 @@ bool SockAddr::isLO() const noexcept
 {
     switch(store.sa.sa_family) {
     case AF_INET: return store.in.sin_addr.s_addr==htonl(INADDR_LOOPBACK);
-#ifdef AF_INET6
     case AF_INET6: return IN6_IS_ADDR_LOOPBACK(&store.in6.sin6_addr);
-#endif
     default: return false;
     }
 }
@@ -581,9 +571,7 @@ bool SockAddr::isMCast() const noexcept
 {
     switch(store.sa.sa_family) {
     case AF_INET: return IN_MULTICAST(ntohl(store.in.sin_addr.s_addr));
-#ifdef AF_INET6
     case AF_INET6: return IN6_IS_ADDR_MULTICAST(&store.in6.sin6_addr);
-#endif
     default: return false;
     }
 }
@@ -641,12 +629,10 @@ SockAddr SockAddr::any(int af, unsigned port)
         ret->in.sin_addr.s_addr = htonl(INADDR_ANY);
         ret->in.sin_port = htons(port);
         break;
-#ifdef AF_INET6
     case AF_INET6:
         ret->in6.sin6_addr = IN6ADDR_ANY_INIT;
         ret->in6.sin6_port = htons(port);
         break;
-#endif
     default:
         throw std::invalid_argument("Unsupported address family");
     }
@@ -661,12 +647,10 @@ SockAddr SockAddr::loopback(int af, unsigned port)
         ret->in.sin_addr.s_addr = htonl(INADDR_LOOPBACK);
         ret->in.sin_port = htons(port);
         break;
-#ifdef AF_INET6
     case AF_INET6:
         ret->in6.sin6_addr = IN6ADDR_LOOPBACK_INIT;
         ret->in6.sin6_port = htons(port);
         break;
-#endif
     default:
         throw std::invalid_argument("Unsupported address family");
     }
@@ -688,7 +672,6 @@ std::ostream& operator<<(std::ostream& strm, const SockAddr& addr)
             strm<<':'<<ntohs(addr->in.sin_port);
         break;
     }
-#ifdef AF_INET6
     case AF_INET6: {
             char buf[INET6_ADDRSTRLEN+1];
             if(evutil_inet_ntop(AF_INET6, &addr->in6.sin6_addr, buf, sizeof(buf))) {
@@ -704,7 +687,6 @@ std::ostream& operator<<(std::ostream& strm, const SockAddr& addr)
                 strm<<':'<<port;
             break;
     }
-#endif
     case AF_UNSPEC:
         strm<<"<>";
         break;