+++ tools/pvxvct.cpp
@@ -28,6 +28,7 @@
 #include <pvaproto.h>
 
 namespace pva = pvxs;
+using pvxs::impl::SB;
 namespace {
 
 DEFINE_LOGGER(out, "pvxvct");
@@ -90,7 +91,8 @@ void usage(const char *name)
                "  -V               Print version and exit.\n"
                "  -C               Show only client Searches\n"
                "  -S               Show only server Beacons\n"
-               "  -B hostip[:port] Listen on the given interface(s).  May be repeated.\n"
+               "  -B <host/ip>[:port] Listen on the given interface(s).  May be repeated.\n"
+               "  -B <mcast>[,ttl#][@iface][:port]\n"
                "  -H host          Show only message sent from this peer.  May be repeated.\n"
                "  -P pvname        Show only searches for this PV name.  May be repeated.\n"
               <<std::endl;
@@ -124,7 +126,7 @@ int main(int argc, char *argv[])
             }
         } opts;
 
-        std::vector<pva::SockAddr> bindaddrs;
+        std::vector<pva::SockEndpoint> bindaddrs;
 
         {
             int opt;
@@ -149,16 +151,8 @@ int main(int argc, char *argv[])
                 case 'S':
                     opts.server = true;
                     break;
-                case 'B': {
-                    pva::SockAddr addr;
-                    int slen = addr.size();
-                    if(evutil_parse_sockaddr_port(optarg, &addr->sa, &slen)) {
-                        throw std::runtime_error(pva::SB()<<"Expected address[:port] to bind.  Not "<<optarg);
-                    }
-                    if(addr.port()==0)
-                        addr.setPort(5076);
-                    bindaddrs.push_back(addr);
-                }
+                case 'B':
+                    bindaddrs.emplace_back(optarg, 5076);
                     break;
                 case 'P':
                     opts.pvnames.insert(optarg);
@@ -206,20 +200,19 @@ int main(int argc, char *argv[])
 
         auto searchCB = [&opts](const pva::UDPManager::Search& msg)
         {
-            if(!opts.client || !opts.allowPeer(msg.src))
+            if(!opts.client || !opts.allowPeer(msg.origSrc))
                 return;
 
             if(!opts.pvnames.empty()) {
                 bool show = false;
                 for(const auto pv : msg.names) {
-                    show = opts.pvnames.find(pv.name)!=opts.pvnames.end();
-                    if((show = opts.pvnames.find(pv.name)!=opts.pvnames.end()))
+                    if((show = (opts.pvnames.find(pv.name)!=opts.pvnames.end())))
                         break;
                 }
                 if(!show)
                     return;
             }
-            log_info_printf(out, "%s Searching for:\n", msg.src.tostring().c_str());
+            log_info_printf(out, "%s Searching for:\n", msg.origSrc.tostring().c_str());
             for(const auto pv : msg.names) {
                 log_info_printf(out, "  \"%s\"\n", pv.name);
             }
@@ -245,7 +238,7 @@ int main(int argc, char *argv[])
                                    manager.onBeacon(baddr, beaconCB));
             listeners.back().first->start();
             listeners.back().second->start();
-            log_debug_printf(out, "Bind: %s\n", baddr.tostring().c_str());
+            log_debug_printf(out, "Bind: %s\n", (SB()<<baddr).str().c_str());
         }
 
 
@@ -261,7 +254,7 @@ int main(int argc, char *argv[])
         return 0;
     }catch(std::runtime_error& e) {
         errlogFlush();
-        std::cerr<<"Error: "<<e.what()<<std::endl;
+        std::cerr<<ERL_ERROR ": "<<e.what()<<std::endl;
         return 1;
     }
 }